Display usernames in the contributors table.
Not all authorized contributors are required to have a first and last name. This
adjusts the "All Contributors" table to display the two things all authorized
contributors must have: a chef username and at least one github account linked.

Note: this is super not cool in the efficiency land. Like it makes a bunch of
queries and that is a bummer.

[skip ci]